refactor(core): restructure les services et interfaces d'authentification et de gestion des CV
- Refactorise et organise les services `AuthService` et `CvService` en fichiers TypeScript modulaires.
- Déplace et regroupe les interfaces liées (`AuthState`, `CvMetadata`, `User`) dans un dossier `core/interfaces`.
- Uniformise les chemins d'importation pour refléter la nouvelle organisation des fichiers.
- Améliore la lisibilité et la maintenabilité du code avec un découpage plus granulaire.
