+++ src/app/core/interfaces/cv.ts
@@ -1,0 +1,16 @@
+export interface CvMetadata {
+  id: string;
+  userId: string;
+  fileName: string;
+  originalName?: string;
+  filePath: string;
+  fileSize: number;
+  mimeType: string;
+  version: string | null;
+  downloadCount: number;
+  createdAt: string;
+}
+
+export interface UploadCvResponse {
+  path: string;
+}
