+++ src/app/core/interfaces/index.ts
@@ -1,0 +1,4 @@
+// Export all core interfaces
+export * from './user';
+export * from './auth';
+export * from './cv';
