+++ src/app/core/interfaces/user.ts
@@ -1,0 +1,7 @@
+export interface User {
+  readonly id: string;
+  readonly email: string;
+  readonly roles: string[];
+  readonly createdAt: string;
+  readonly updatedAt: string;
+}
